NAS-123925: Moved classes
